third_party/blink/web_tests/platform/generic/fast/js/null-char-in-string-expected.txt

This test checks that null characters are allowed in JavaScript strings, rather than causing a parse error.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS String("\0").length is 1
PASS successfullyParsed is true

TEST COMPLETE
